Adds "BatchNormInference" construct to Glenside, as a hack
We will be bringing batch norms into Glenside as these BatchNormInference nodes,
rather than running SimplifyInference in Relay and bringing them in as
adds/multiplies/divides/etc. This commit adds support for this new construct,
and parses it from Relay. I've added a flag to the Relay-to-Glenside compiler to
try and call this out specifically as a hack.

In the future, ideally, we'd bring in either batch norm nodes, or we'd bring in
the simplified components of batch norm inference and do efficient codegen for
them.

See #55