Helper returns Option, so we can opt out of generating code
This helps us out because then we don't haave to worry what ends up on the
worklist. If something is on the worklist that we don't need to generate code
for, we can just ignore it by returning None. Of course, this will cause errors
if someone tries to look up the result of codegen for that specific eclass Id!