Add macro for relay-to-glenside tests, enable it for conv2d
This change copies the test! macro from the from_relay mod
and modifies it for use for the new relay-to-glenside-via-
rewrites pipeline. It also modifies from_relay for conv2d
so that it's able to be tested using the new test! macro.
Specifically, we make it so that from_relay simply constructs
a RelayOperatorCall for conv2d, which is then rewritten by
the rewrite.