from_relay no longer returns tuples of (Id, Option<Id>); just returns Id
Originally, Mike was returning these tuples because from_relay would add
both the glenside AND the relay versions of the expressions into the
egraph, and later we'd manually union them. Now, this all works via
rewrites, and the rewrites do the unions automatically. So we return
from_relay back to its original API.